move music and soundpool to audio and add documentation
