+++ library/src/cpp/audio/renderable_audio.hpp
@@ -2,9 +2,11 @@
 
 #include <cstdint>
 
-/// Interface for writing audio into raw audio stream
+/// Interface for writing audio into raw audio stream.
 class renderable_audio {
 public:
-    virtual void render(int16_t *, uint32_t) = 0;
+    /// Render this audio into the stream for num_frames of frames.
+    virtual void render(int16_t* stream, uint32_t num_frames) = 0;
+protected:
     virtual ~renderable_audio() = default;
 };
